server/requirements.txt
fastapi==0.98.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
uvicorn==0.20.0
boto3==1.26.55
pre-commit==2.17.0
python-dotenv==1.0.0
requests==2.31.0
cryptography==39.0.2
stripe==2.49.0
jinja2==3.1.2
bcrypt==4.0.1
